requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 26.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
asttokens=3.0.1=pyhd8ed1ab_0
attrs=26.1.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
bottleneck=1.4.2=py310haa0f9ac_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2026.2.25=hbd8a1cb_0
comm=0.2.3=pyhe01879c_0
configargparse=1.7.5=pypi_0
dash=4.1.0=pypi_0
debugpy=1.8.20=py310h25320af_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.3.1=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
fastjsonschema=2.21.2=pypi_0
flask=3.1.3=pypi_0
icu=73.1=h6a678d5_0
importlib-metadata=9.0.0=pypi_0
intel-openmp=2025.0.0=h06a4308_1171
ipykernel=7.2.0=pyha191276_1
ipython=8.37.0=pyh8f84b5b_0
ipywidgets=8.1.8=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
joblib=1.5.3=pypi_0
jsonschema=4.26.0=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter_client=8.8.0=pyhcf101f3_0
jupyter_core=5.9.1=pyhc90fa1f_0
jupyterlab-widgets=3.0.16=pypi_0
keyutils=1.6.3=hb9d3cd8_0
krb5=1.22.2=ha1258a1_0
laspy=2.7.0=pypi_0
ld_impl_linux-64=2.44=h9e0c5a2_3
libedit=3.1.20250104=pl5321h7949ede_0
libexpat=2.7.4=h7354ed3_0
libffi=3.4.4=h6a678d5_1
libgcc=15.2.0=h69a1729_7
libgcc-ng=15.2.0=h166f726_7
libgomp=15.2.0=h4751f2c_7
libhwloc=2.12.1=default_hf1bbc79_1000
libnsl=2.0.0=h5eee18b_0
libsodium=1.0.21=h280c20c_3
libstdcxx=15.2.0=h39759b7_7
libstdcxx-ng=15.2.0=hc03a8fd_7
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libxml2=2.13.9=h2c43086_0
libzlib=1.3.1=hb25bd0a_0
lz4=4.4.5=pypi_0
matplotlib-inline=0.2.1=pyhd8ed1ab_0
mkl=2025.0.0=hacee8c2_941
mkl-service=2.5.2=py310hacdc0fc_0
mkl_fft=2.1.1=py310h8fe796d_0
mkl_random=1.3.0=py310h505adc9_0
narwhals=2.18.1=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.5=h7934f7d_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
numexpr=2.14.1=py310h41d4191_0
numpy=2.2.6=pypi_0
open3d=0.19.0=pypi_0
openssl=3.6.1=h35e630c_1
packaging=25.0=py310h06a4308_1
pandas=2.3.3=py310h23c847b_1
parso=0.8.6=pyhcf101f3_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=12.1.1=pypi_0
pip=26.0.1=pyhc872135_0
platformdirs=4.9.4=pyhcf101f3_0
plotly=6.6.0=pypi_0
prompt-toolkit=3.0.52=pyha770c72_0
psutil=7.2.2=py310h139afa4_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=23.0.1=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
python=3.10.20=h741d88c_0
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2025.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2026.1.post1=py310h06a4308_0
pyzmq=27.1.0=py310hc4bea81_2
readline=8.3=hc2a1206_0
referencing=0.37.0=pypi_0
rerun-sdk=0.30.2=pypi_0
retrying=1.4.2=pypi_0
rosbags=0.11.0=pypi_0
rpds-py=0.30.0=pypi_0
ruamel-yaml=0.19.1=pypi_0
scikit-learn=1.7.2=pypi_0
setuptools=80.10.2=py310h06a4308_0
six=1.17.0=py310h06a4308_0
sqlite=3.51.2=h3e8d24a_0
stack_data=0.6.3=pyhd8ed1ab_1
tbb=2022.3.0=h698db13_0
tbb-devel=2022.3.0=h698db13_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
tornado=6.5.5=py310h7c4b9e2_0
traitlets=5.14.3=pyhd8ed1ab_1
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2026a=he532380_0
wcwidth=0.6.0=pyhd8ed1ab_0
wheel=0.46.3=py310h06a4308_0
widgetsnbextension=4.0.15=pypi_0
xarray=2025.4.0=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.8.2=h448239c_0
zeromq=4.3.5=h41580af_10
zipp=3.23.0=pypi_0
zlib=1.3.1=hb25bd0a_0
zstandard=0.25.0=pypi_0
